Move iterators for decoding spatial differencing data to the appropriate module
